Limit shared_ptr_deleter's GIL acquisition to clients with multithreading enabled & a "new enough" Python version.
Only compile in this behavior for Python 2.4 & newer (because
PyEval_ThreadsInitialized was added in 2.4) and only if Python was compiled
with threading support (WITH_THREAD) so that Boost Python continues to work
as-is for other users.

Only acquire the GIL if threading support has been initialized
at runtime to help avoid overhead in clients that don't require it.