Reduce smart_graph overhead when thousands of casts are registered.
The original design of smart_graph featured a dense N^2 table of distances,
which allowed fast lookup using a simple offset.  This works really well when
N is small.  However, in a system with 4000 entries (which is not a
hypothetical number) smart_graph consumes 122MB.  Additionally, all entries in
the table are initialized each time the table is resized.  In a process where
a module is imported, some work is done, then another module is imported,
etc..., the table may have to be re-initialized many times as additional casts
are introduced, which becomes a significant drag[1] on the time required to
import Boost.Python libraries.

This change addresses both the storage requirements and re-initialization
costs.  To address the memory use, smart_graph holds an unordered_map from
(src, dst) vertex pairs to (distance, version) pairs.  While this
unfortunately means that memory consumption will increase for very small
modules, as the number of vertices increases, the additional memory overhead
required for hash table entries should be compensated for by the sparseness[2]
of the graph.  Similarly, the performance loss from hash table lookups is
negligible compared to the savings from avoiding dense initialization.  Also,
there is a outer level of caching (see cache_t) that helps hide the increased
smart_graph costs from clients making repeated queries over the same casts.

Sparse re-initialization is implemented by tagging each distance table
entry with a version.  This corresponds to the version of the graph
from which the distance was computed.  As the graph grows, its version
is incremented.  When an existing entry is found, its version can be checked
against the graph's current version to determine if it is still valid or
should be re-computed.

[1] Roughly 250ms is spent just filling the tables with the unreachable
    sentinel value during a startup sequence that registers around 4000 types
    across dozens of libraries.

[2] In internal testing, the densest observed graphs have ~3700 vertices but
    only 1100 initialized entries in the smart_graph.  The dense
    implementation consumes 104MB while the sparse table uses only 37KB.