Ensure that num_edges() runs in constant time.
The default type of adjacency_list's EdgeList (not to be confused with
OutEdgeList) is listS (a.k.a std::list.)  Many versions of libstdc++ implement
std::list::size in O(n) time.  Switch EdgeList to vecS for substantial
speedups when adding casts to the inheritance graph.

(While C++ 11 requires std::list::size() be implemented in constant time, the
contiguous layout of vector should still be advantageous, especially as
edges are never removed.)