Pass alignment to holder allocator
aca3c8 added an alignment argument to holder::allocate which defaults to 1.  However, make_constructor.hpp uses this function to create storage for a pointer and was never changed to pass an alignment.

This causes unaligned pointers to be created.  This can be detected as a failure by running projects under Clang ubsan.