Fix tutorial documentation Jamroot/Jamfile references
The tutorial was incorrectly stating that the tutorial directory contains
a Jamroot file, when it actually contains a Jamfile. The Jamroot is in
the parent example directory. This was causing confusion as the link
pointed to example/Jamroot but the text suggested it was in the tutorial
directory.

Updated the documentation to clarify:
- The tutorial directory contains hello.cpp, hello.py, and Jamfile
- The parent example directory contains the Jamroot template
- Made the link description clearer about the file location

Fixes #1045

Signed-off-by: Samaresh Kumar Singh <ssam3003@gmail.com>